chore(payments-paypal): move getCheckoutToken to CheckoutTokenManager
