+++ libs/payments/paypal/src/index.ts
@@ -4,6 +4,7 @@
 
 export * from './lib/paypalCustomer/paypalCustomer.manager';
 export * from './lib/paypalCustomer/paypalCustomer.factories';
+export * from './lib/checkoutToken.manager';
 export * from './lib/paypal.client';
 export * from './lib/paypal.client.config';
 export * from './lib/paypal.client.types';
