+++ libs/payments/paypal/src/lib/paypal.manager.spec.ts
@@ -20,7 +20,6 @@ import { MockAccountDatabaseNestFactory } from '@fxa/shared/db/mysql/account';
 import {
   NVPCreateBillingAgreementResponseFactory,
   NVPBAUpdateTransactionResponseFactory,
-  NVPSetExpressCheckoutResponseFactory,
 } from './factories';
 import { PayPalClient } from './paypal.client';
 import { PayPalManager } from './paypal.manager';
@@ -349,27 +348,6 @@ describe('PayPalManager', () => {
       mockCustomer.id
     );
     expect(result).toEqual([]);
-  });
-
-  describe('getCheckoutToken', () => {
-    it('returns token and calls setExpressCheckout with passed options', async () => {
-      const currencyCode = faker.finance.currencyCode();
-      const token = faker.string.uuid();
-      const successfulSetExpressCheckoutResponse =
-        NVPSetExpressCheckoutResponseFactory({
-          TOKEN: token,
-        });
-
-      jest
-        .spyOn(paypalClient, 'setExpressCheckout')
-        .mockResolvedValue(successfulSetExpressCheckoutResponse);
-
-      const result = await paypalManager.getCheckoutToken(currencyCode);
-
-      expect(result).toEqual(successfulSetExpressCheckoutResponse.TOKEN);
-      expect(paypalClient.setExpressCheckout).toBeCalledTimes(1);
-      expect(paypalClient.setExpressCheckout).toBeCalledWith({ currencyCode });
-    });
   });
 
   describe('processZeroInvoice', () => {
